refactor: move all workday api calls to designated folder
